pyproject.toml
[project]
name = "AtherisLiteLLM"
version = "0.2.013"
description = "AI-powered Python fuzzer using LiteLLM and Atheris to automatically generate and execute fuzzing harnesses."
authors = [
  { name="Mario Marku", email="mariomarku7@gmail.com" }
]
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.13,"

dependencies = [
  "litellm>=1.0.0",
  "pyyaml>=6.0",
  "atheris>=2.3.0",
  "radon>=5.0.0",
  "requests>=2.31.0",
  "tenacity>=9.1.4"
]

[project.urls]
Homepage = "https://github.com/mariobx"
Repository = "https://github.com/mariobx/AtherisLiteLLM"

[project.scripts]
atherislitellm = "ai_fuzzer.cli:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
